// generated/src/app/api-models/stage-auth-prov-support.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class StageAuthProvSupport {
  batchId: string;
  transactionId: number;
  seqNo: number;
  lastName: string;
  firstName: string;
  seqProvId: number;
  seqProvContract: number;
  prevSeqProvId: number;
  seqProvxId: number;
  seqCovProvGrp: number;
  inAreaServiceStatus: string;
}